csv files used in NSFG_From_SAS.Rmd
These are the files that are run in the NSFG_From_SAS.Rmd code. Basically they're a subset of interesting variables from the SAS dataset, but in a file size that's program agnostic (and small enough to be loaded onto GitHub). They maintain the original variable names from the 2015-2017 NSFG datasets and are the closest to the "raw data" downloaded from the NSFG website